jabber: Don't pass a readonly empty string to imcb_chat_topic()
Fixes crashes when a topic is unset and strip_html is set to always.
Turns out that the strip_html() function does a strcpy at the end which
may write a single null byte over the null byte of the empty string, and
if it's not in writable memory, that blows up.

Thanks to iamthemcmaster / mcm for pointing this out